a much needed cleanup
