--- __init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-__author__ = 'Will Brennan'
-
-
-# Built-in Modules
-# Standard Modules
-# Custom Modules
-from main import *
-import scripts
